refactor(agent): extract lifecycle/sync subsystem into mixin holder
Move start() boot orchestration, random-sampling prover wiring, peer/CG
sync (warm-core, catchup, paged fetch, sync-verify worker), subscription
state bookkeeping and shared-memory TTL cleanup (~3.2k lines) into
dkg-agent-lifecycle.ts. Relocate AGENT_SYSTEM_GRAPH to DKGAgentBase.
